test: msw 핸들러 수정
