+++ frontend/src/mocks/handlers/progressesHandlers.ts
@@ -13,12 +13,12 @@ export const progressesHandlers = [
     if (requestAuthToken === newAccessToken && memberId === '1')
       return res(
         // 성공
-        ctx.status(200),
-        ctx.json({ progresses: [{ progressId: 1, nickname: '하루', currentCycle: 1, step: 'planning' }] }),
+        // ctx.status(200),
+        // ctx.json({ progresses: [{ progressId: 1, nickname: '하루', currentCycle: 1, step: 'planning' }] }),
 
         // progresses가 없는 경우
-        // ctx.status(200),
-        // ctx.json({ progresses: null }),
+        ctx.status(200),
+        ctx.json({ progresses: null }),
 
         // 그 외 에러
         // ctx.status(404),
@@ -57,16 +57,50 @@ export const progressesHandlers = [
 
     return res(
       // 성공
-      ctx.status(200),
-      ctx.json({ progresses: [{ progressId: 1, nickname: '하루', currentCycle: 1, step: 'planning' }] }),
+      // ctx.status(200),
+      // ctx.json({ progresses: [{ progressId: 1, nickname: '하루', currentCycle: 1, step: 'planning' }] }),
 
       // progresses가 없는 경우
-      // ctx.status(200),
-      // ctx.json({ progresses: null }),
+      ctx.status(200),
+      ctx.json({ progresses: null }),
 
       // 그 외 에러
       // ctx.status(404),
       // ctx.json({ code: 1234, message: '에러' }),
+
+      ctx.delay(1000),
+    );
+  }),
+
+  rest.post('/api/studies/:studyId/progresses', (req, res, ctx) => {
+    const accessToken =
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNjkxNTY4NDI4LCJleHAiOjE2OTE1NzIwMjh9.BfGH7jBxO_iixmlpzxHKV7d9ekJPegLxrpY9ME066ro';
+    const requestAuthToken = req.headers.get('Authorization')?.split(' ')[1];
+    const newAccessToken =
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxMjM0NTY3fQ.NUiutjXo0mcIBU5fWxfjpBEvPxakFiBaUCg4THKAYpQ';
+
+    if (requestAuthToken === newAccessToken) {
+      return res(
+        // 성공
+        ctx.status(201),
+
+        // 실패
+        // ctx.status(404),
+        // ctx.json({ message: '에러', code: 1234 }),
+
+        ctx.delay(1000),
+      );
+    }
+
+    if (accessToken !== requestAuthToken) return res(ctx.status(401), ctx.delay(1000));
+
+    return res(
+      // 성공
+      ctx.status(201),
+
+      // 실패
+      // ctx.status(404),
+      // ctx.json({ message: '에러', code: 1234 }),
 
       ctx.delay(1000),
     );
